+++ library/multichoicerecyclerview/src/main/java/com/davidecirillo/multichoicerecyclerview/MultiChoiceToolbar.java
@@ -60,37 +60,47 @@ public class MultiChoiceToolbar {
             this.mToolbar = toolbar;
         }
 
-        public Builder setDefaultPrimaryColor(int defaultPrimaryColor) {
+        /**
+         * Set the colours when the toolbar is the default one
+         *
+         * @param defaultPrimaryColor     The color used for the background of the toolbar when is in default mode
+         * @param defaultPrimaryColorDark The color used for the status bar background when is in default mode
+         * @return Builder so you can chain together setters and build
+         */
+        public Builder setDefaultColours(int defaultPrimaryColor, int defaultPrimaryColorDark) {
             this.mDefaultPrimaryColor = defaultPrimaryColor;
-            return this;
-        }
-
-        public Builder setDefaultToolbarTitle(String defaultToolbarTitle) {
-            this.mDefaultToolbarTitle = defaultToolbarTitle;
-            return this;
-        }
-
-        public Builder setSelectedToolbarTitle(String selectedToolbarTitle) {
-            this.mSelectedToolbarTitle = selectedToolbarTitle;
-            return this;
-        }
-
-        public Builder setDefaultPrimaryColorDark(int defaultPrimaryColorDark) {
             this.mDefaultPrimaryColorDark = defaultPrimaryColorDark;
             return this;
         }
 
-        public Builder setMultiPrimaryColor(int multiPrimaryColor) {
-            this.mMultiPrimaryColor = multiPrimaryColor;
-            return this;
-        }
-
-        public Builder setMultiPrimaryColorDark(int multiPrimaryColorDark) {
+        /**
+         * Set the colours when the toolbar is the multi choice one
+         *
+         * @param multiPrimaryColor     The color used for the background of the toolbar when is in multi choice mode
+         * @param multiPrimaryColorDark The color used for the status bar background when is in multi choice mode
+         * @return Builder so you can chain together setters and build
+         */
+        public Builder setMultiChoiceColours(int multiPrimaryColor, int multiPrimaryColorDark) {
+            this.mMultiPrimaryColorDark = multiPrimaryColorDark;
             this.mMultiPrimaryColorDark = multiPrimaryColorDark;
             return this;
         }
 
-        public Builder setIcon(int icon, View.OnClickListener action) {
+        /**
+         * Set the titles for the differente state of the toolbar default/multiChoice
+         *
+         * @param defaultTitle  Title when the toolbar shown is the default one
+         * @param selectedTitle Title shown when some item are selected. Will be of the format "itemCount title" where "itemCount" is the number of
+         *                      selected item and "title" is this param
+         * @return Builder so you can chain together setters and build
+         */
+        public Builder setTitles(String defaultTitle, String selectedTitle) {
+            this.mDefaultToolbarTitle = defaultTitle;
+            this.mSelectedToolbarTitle = selectedTitle;
+            return this;
+        }
+
+        public Builder setDefaultIcon(int icon, View.OnClickListener action) {
             mIcon = icon;
             mIconAction = action;
             return this;
@@ -115,7 +125,7 @@ public class MultiChoiceToolbar {
 
     /* Getters */
 
-    public AppCompatActivity getAppCompatActivity() {
+    AppCompatActivity getAppCompatActivity() {
         return mAppCompatActivity;
     }
 
@@ -123,27 +133,27 @@ public class MultiChoiceToolbar {
         return mToolbar;
     }
 
-    public String getDefaultToolbarTitle() {
+    String getDefaultToolbarTitle() {
         return mDefaultToolbarTitle;
     }
 
-    public String getSelectedToolbarTitle() {
+    String getSelectedToolbarTitle() {
         return mSelectedToolbarTitle;
     }
 
-    public int getDefaultPrimaryColor() {
+    int getDefaultPrimaryColor() {
         return mDefaultPrimaryColor;
     }
 
-    public int getDefaultPrimaryColorDark() {
+    int getDefaultPrimaryColorDark() {
         return mDefaultPrimaryColorDark;
     }
 
-    public int getMultiPrimaryColor() {
+    int getMultiPrimaryColor() {
         return mMultiPrimaryColor;
     }
 
-    public int getMultiPrimaryColorDark() {
+    int getMultiPrimaryColorDark() {
         return mMultiPrimaryColorDark;
     }
 
@@ -151,11 +161,11 @@ public class MultiChoiceToolbar {
         return mIcon;
     }
 
-    public View.OnClickListener getIconAction() {
+    View.OnClickListener getIconAction() {
         return mIconAction;
     }
 
-    public Listener getToolbarListener() {
+    Listener getToolbarListener() {
         return mListener;
     }