+++ CHANGELOG.md
@@ -5,11 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Added
 ### Changed
 - Moved logic related to the MultiChoiceRecyclerView class into the MultiChoiceAdapter
 - Internal improvements getting rid of holding views states
-### Fixed
+- Aggregated some of the method of the MultiChoiceToolbar Builder like titles, defaultColors, selectedColors and change "setIcon" to "setDefaultIcon"
+
 ### Deleted
 - MultiChoiceRecyclerView class